// src/lib/types/api.types.ts
export interface CreateDeckBody {
  title: string;
  description?: string;
  is_public?: boolean;
  primary_color_hex?: string;
}